Add ListenAndServe to dmsghttp
The commit removes func serveHTTPOverDmsg from cmd/dmsg-discovery and instead creates a common public func ListenAndServe in dmsghttp so that any service can use it.
